+ plots, data view, requested slides, python notebook, csv
the "_population_plot.html" files are exported using plotly and can be displayed directly on a webpage. They can be embedded to show our data.

The data view has been edited to display the data plots as well as allow navigation between them. Feel free to change up the visuals from here if what i've placed is ugly.

The python notebooks contain the code I used to create the datasets, process, and merge the datasets, which are uploaded in a merged form in the .csv files